Making it easier to import and use
