+++ packages/molly/.env.example
@@ -1,0 +1,2 @@
+OPENAI_API_KEY="https://platform.openai.com/account/api-keys"
+OPENAI_TOKEN_LIMIT=4000
