+++ packages/molly/package.json
@@ -39,12 +39,13 @@
 		"eslint": "^8.28.0",
 		"eslint-config-prettier": "^8.5.0",
 		"eslint-plugin-svelte3": "^4.0.0",
-		"openai": "^3.2.1",
-		"mdsvex": "^0.10.6",
 		"gpt-tokenizer": "^1.0.4",
+		"mdsvex": "^0.10.6",
+		"openai": "^3.2.1",
 		"prettier": "^2.8.0",
 		"prettier-plugin-svelte": "^2.8.1",
 		"publint": "^0.1.9",
+		"sse.js": "^0.6.1",
 		"svelte": "^3.54.0",
 		"svelte-check": "^3.0.1",
 		"tslib": "^2.4.1",
