+++ packages/molly/src/lib/server/openai/index.ts
@@ -2,4 +2,5 @@ export * from './chat';
 export * from "./moderate";
 
 import prompts from './prompts';
-export { prompts };
+import methods from './methods';
+export { prompts, methods };
