+++ packages/molly/src/lib/server/openai/methods/index.ts
@@ -1,0 +1,9 @@
+import { chatCompletionPost } from "./post";
+
+export default {
+  chat: {
+    completion: {
+      POST: chatCompletionPost
+    }
+  }
+}
