+++ packages/molly/src/routes/+server.ts
@@ -1,0 +1,3 @@
+import { methods } from '$lib/server/openai';
+
+export const POST = methods.chat.completion.POST;
